@@ -3,11 +3,11 @@ description: >2-
   Package for creating canvas animations based on time and math functions.
 
   Fun + canvas -> funvas :)
-version: 0.0.3+1
+version: 0.0.4-nullsafety.1
 repository: https://github.com/creativecreatorormaybenot/funvas
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0-29.10.beta <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
