@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:golden_toolkit/golden_toolkit.dart';
 
-Future<void> main(FutureOr<void> Function() testMain) async {
+Future<void> testExecutable(FutureOr<void> Function() testMain) async {
   await loadAppFonts();
   await testMain();
 }
